Fixed cursor on resource list.
Due to a incredibly stupid mistake on my part, the cursor would always
glitch out randomly from the regular windows cursor to the "SizeWS"
whilst hovering over the resource list. Not only did it not work well at
ALL, but I typed up a bunch of junky, un-optimized code to get it to
"work". I fixed the glitch and removed the un-necessary code, so now the
cursor transforms as it should perfectly smoothly, and there's no extra
"junk code" to make it work. It just does. :smile: