Added ability to test game in LOVE 2D
Added the basic ability to generate code and test the game you just made
in LOVE 2D! As of now, it only works with "Move" actions, and your game
errors if you have no objects/move actions. However, this is all
temporary and will be fixed.

Now includes 7-Zip by Igor Pavlov, which the program uses for zipping up
.love files. The license file had to be updated to include it.