Added new logos created generously by a contributor!
I don't know if he wants his real name mentioned or not, which is the only reason it's not included. If so, well, I'm really sorry about that! :( Just figured it'd be better to go un-credited than to have your identity revealed in the case that you wanted to "stay undercover." :|